Global changes in the structure
