+++ MyWebRecruit.Services/MyWebRecruit.Services.csproj
@@ -6,7 +6,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\MyWebRecruit.Data\MyWebRecruit.Data.csproj" />
-    <ProjectReference Include="..\MyWebRecruit.Domain\MyWebRecruit.Domain.csproj" />
   </ItemGroup>
 
 </Project>
